refactor(dasboard): refine add athletes sheet & implement virtual scroll
