+++ src/features/dashboard/tournament/builder/components/groups-tab/index.tsx
@@ -3,8 +3,8 @@ import { DndContext } from '@dnd-kit/core';
 import { toast } from 'sonner';
 import { useBuilderManagerQuery } from '../../hooks/use-builder-manager-query';
 import { AddGroupDialog } from '../dialogs/add-group-dialog';
-import { AddAthletesSheet } from '../dialogs/add-athletes-sheet';
 import { GroupSettingsSheet } from '../dialogs/group-settings-sheet';
+import { AddAthletesSheet } from './add-athletes-sheet';
 import { AthletePool } from './athlete-pool';
 import { GroupRosterTable } from './group-roster-table';
 import { GroupsRail } from './groups-rail';
